AIModels/fuckaround/requirements.txt
streamlit
langchain
langchain-google-genai
google-generativeai
tiktoken
python-dotenv
PyPDF2
